Use a local fn
